+++ examples/plot_sleep_staging_usleep.py
@@ -149,12 +149,12 @@ valid_set = splitted['valid']
 # the memory requirements for this example we use shorter sequences of 3
 # windows.
 
-from braindecode.samplers import SequenceSampler, USleepSampler
+from braindecode.samplers import SequenceSampler
 
-n_windows = 10  # Sequences of 3 consecutive windows; originally 35 in paper
-n_windows_stride = 10  # Non-overlapping sequences
+n_windows = 3  # Sequences of 3 consecutive windows; originally 35 in paper
+n_windows_stride = 3  # Non-overlapping sequences
 
-train_sampler = USleepSampler(train_set.get_metadata(), n_windows, n_windows_stride)
+train_sampler = SequenceSampler(train_set.get_metadata(), n_windows, n_windows_stride)
 valid_sampler = SequenceSampler(valid_set.get_metadata(), n_windows, n_windows_stride)
 
 # Print number of examples per class