+++ docs/whats_new.rst
@@ -55,7 +55,8 @@ Enhancements
 - Only load data if needed during preprocessing (e.g., allow timecrop without loading) (:gh:`164` by `Robin Tibor Schirrmeister`_)
 - Adding option to sort filtered channels by frequency band for the filterbank in :func:`braindecode.datautil.filterbank` (:gh:`185` by `Lukas Gemein`_)
 - Adding the USleep model :class:`braindecode.models.USleep` (:gh:`282` by `Theo Gnassounou`_ and `Omar Chehab`_)
-
+- Adding the USleep sampler :class:`braindecode.samplers` (:gh:`` by `Theo Gnassounou`_)
+  
 Bugs
 ~~~~
 - Amplitude gradients are correctly computed for layers with multiple filters